backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
openai>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0
itsdangerous>=2.1.0
